+++ app/domain/authentication/authn_jwt/vendor_configurations/configuration_jwt_generic_vendor.rb
@@ -2,8 +2,9 @@ module Authentication
   module AuthnJwt
     module VendorConfigurations
       # Mock JWTConfiguration class to use it to develop other part in the jwt authenticator
-      class ConfigurationJWTGenericVendor < ConfigurationInterface
-
+      class ConfigurationJWTGenericVendor
+        # These are dependencies in class integrating different parts of the jwt authentication
+        # :reek:CountKeywordArgs
         def initialize(
           authenticator_input:,
           logger: Rails.logger,
@@ -12,8 +13,9 @@ module Authentication
           validate_and_decode_token_class:  Authentication::AuthnJwt::ValidateAndDecode::ValidateAndDecodeToken,
           validate_resource_restrictions_class: Authentication::ResourceRestrictions::ValidateResourceRestrictions,
           extract_token_from_credentials: Authentication::AuthnJwt::InputValidation::ExtractTokenFromCredentials.new,
-          create_identity_provider: Authentication::AuthnJwt::IdentityProviders::CreateIdentityProvider.new
-
+          create_identity_provider: Authentication::AuthnJwt::IdentityProviders::CreateIdentityProvider.new,
+          create_constraints: Authentication::AuthnJwt::RestrictionValidation::CreateConstrains.new,
+          fetch_mapping_claims: Authentication::AuthnJwt::RestrictionValidation::FetchMappingClaims.new
         )
           @logger = logger
           @authentication_parameters_class = authentication_parameters_class
@@ -22,6 +24,8 @@ module Authentication
           @validate_resource_restrictions_class = validate_resource_restrictions_class
           @extract_token_from_credentials = extract_token_from_credentials
           @create_identity_provider = create_identity_provider
+          @create_constraints = create_constraints
+          @fetch_mapping_claims = fetch_mapping_claims
 
           @logger.debug(LogMessages::Authentication::AuthnJwt::CreatingAuthenticationParametersObject.new)
           @authentication_parameters = @authentication_parameters_class.new(
@@ -42,7 +46,8 @@ module Authentication
             role_name: jwt_identity,
             constraints: constraints,
             authentication_request: @restriction_validator_class.new(
-              decoded_token: @authentication_parameters.decoded_token
+              decoded_token: @authentication_parameters.decoded_token,
+              mapped_claims: mapped_claims
             )
           )
         end
@@ -60,6 +65,10 @@ module Authentication
           @extract_token_from_credentials.call(
             credentials: authenticator_input.request.body.read
           )
+        end
+
+        def mapped_claims
+          @mapped_claims ||= @fetch_mapping_claims.call(authentication_parameters: @authentication_parameters)
         end
 
         def jwt_identity_from_request
@@ -129,8 +138,8 @@ module Authentication
         end
 
         def constraints
-          @constraints ||= Authentication::Constraints::MultipleConstraint.new(
-            Authentication::Constraints::NotEmptyConstraint.new
+          @constraints ||= @create_constraints.call(
+            authentication_parameters: @authentication_parameters
           )
         end